Fixed a bug with the wrong type
The writing of the PropertyDefinition was writing an Int for a Revision but reading a Short. Standardized on Short

+++ model/common-model.standard/src/main/java/com/diamondq/common/model/generic/GenericPropertyDefinition.java
@@ -634,7 +634,7 @@ public class GenericPropertyDefinition implements PropertyDefinition {
 
       /* Revision */
 
-      buffer.putInt(1);
+      buffer.putShort((short) 1);
 
       /* mName */
       buffer.putInt(nameBytes.length);
